Change inputs to the calculator from Integer to Double

+++ src/main/java/InterestCalculator.java
@@ -1,19 +1,19 @@
 public class InterestCalculator {
 
-    private static final String INPUT_NOT_PRESENT_ERROR = "One of the inputs is null - principal: %d, percent interest: %d, time: %d";
+    private static final String INPUT_NOT_PRESENT_ERROR = "One of the inputs is null - principal: %f, percent interest: %f, time: %f";
 
-    public Integer calculate(Integer principal, Integer percentageInterest, Integer timeInYears) throws RequiredInputNotPresentException {
+    public double calculate(Double principal, Double percentageInterest, Double timeInYears) throws RequiredInputNotPresentException {
         if (requiredInputsPresent(principal, percentageInterest, timeInYears)) {
             return interestFor(principal, percentageInterest, timeInYears);
         }
         throw new RequiredInputNotPresentException(String.format(INPUT_NOT_PRESENT_ERROR, principal, percentageInterest, timeInYears));
     }
 
-    private Integer interestFor(Integer principal, Integer percentageInterest, Integer timeInYears) {
+    private double interestFor(Double principal, Double percentageInterest, Double timeInYears) {
         return (principal * percentageInterest * timeInYears) / 100;
     }
 
-    private boolean requiredInputsPresent(Integer principal, Integer percentageInterest, Integer timeInYears) {
+    private boolean requiredInputsPresent(Double principal, Double percentageInterest, Double timeInYears) {
         return principal != null && percentageInterest != null && timeInYears != null;
     }
 }
